Add a regression test "suite"
While fiddling with various pieces of the algorithms it's nice to know you haven't broken anything, and which situations have changed outputs. This adds a really lightweight test suite that ensures the APIs always work, and that changes to what arrows are being rendered at least require a snapshot update. It doesn't enforce that the arrows are correct or look perfect, but it at least arms developers with the knowledge of what they are changing so they can inspect it or fix it if it's unexpected.